Add files via upload
The data of test csvs are from the course EESA01, while the conditions for the if statements are from web, the codes are our own. 
The csvs are for testing, you can come up with more csvs to check or just change the data(make sure the header have not been changed)
Make sure to put them in the same folder and change the file name each time, when testing multiple times, remember to delete the data added previously for the max, min, etc. to get the correct result. 
The generated report is a sample for the July data.  If you want to test, start from a blank txt or no txt.